refactor: Refactored code
